+++ src/ui/atoms/Button.tsx
@@ -4,6 +4,7 @@ import styled from '@emotion/styled'
 
 import Spin from './Spin'
 import getBordersStyles, { BorderTypes } from '../../utils/getBordersStyles'
+import { UIElementSizesEnum } from '../ts/enums/UIElementSizesEnum'
 
 enum HtmlType {
   button = 'button',
@@ -21,19 +22,13 @@ enum ButtonType {
   link = 'link',
 }
 
-enum ButtonSizes {
-  small = 'small',
-  medium = 'medium',
-  large = 'large',
-}
-
 
 interface Props {
   children: React.ReactNode
   onClick?: () => void
   htmlType?: HtmlType
   type?: ButtonType
-  size?: ButtonSizes
+  size?: UIElementSizesEnum
   disabled?: boolean
   outlined?: boolean | BorderTypes
   loading?: boolean
@@ -43,7 +38,7 @@ interface Props {
 
 const defaultProps: Partial<Props> = {
   type: ButtonType.primary,
-  size: ButtonSizes.medium,
+  size: UIElementSizesEnum.Medium,
   htmlType: HtmlType.button,
   onClick() { },
   loading: false,
@@ -58,9 +53,16 @@ function Button({
   loading,
   ...restOfProps
 }: Props) {
+
   return (
     <button {...restOfProps} type={htmlType} disabled={disabled || loading}>
-      {loading ? <Spin type={[ButtonType.white, ButtonType.link].includes(type) || outlined ? ButtonType.primary : ButtonType.white} size={20} /> : children}
+      {loading ? (
+        <Spin
+          type={outlined ? type : ButtonType.ghost}
+          size={20}
+        />
+      )
+        : children}
     </button>
   )
 }
@@ -80,24 +82,36 @@ const WrapperButton = styled(Button)`
     ${({ width }) => width && `width: ${width};`}
 
     ${({ size }) => {
-    if (size === ButtonSizes.small)
+    if (size === UIElementSizesEnum.Small)
       return `
         height: 35px;
         font-size: 12px;
         border-radius: 7px;
+
+        @media screen and (min-width: 768px) {
+          font-size: 14px;
+        }
       `
 
-    if (size === ButtonSizes.medium)
+    if (size === UIElementSizesEnum.Medium)
       return `
         height: 44px;
         font-size: 14px;
         border-radius: 7px;
+
+        @media screen and (min-width: 768px) {
+          font-size: 16px;
+        }
       `
 
-    if (size === ButtonSizes.large)
+    if (size === UIElementSizesEnum.Large)
       return `
         height: 50px;
         font-size: 18px;
+
+        @media screen and (min-width: 768px) {
+          font-size: 20px;
+        }
       `
   }}
 
@@ -166,9 +180,5 @@ const WrapperButton = styled(Button)`
 `
 
 WrapperButton.defaultProps = defaultProps
-
-Button.htmlType = HtmlType
-Button.buttonType = ButtonType
-Button.buttonSizes = ButtonSizes
 
 export default WrapperButton