+++ src/ui/atoms/Image.tsx
@@ -1,16 +1,20 @@
 import React from 'react'
 import styled from '@emotion/styled'
 
-type ImageProps = {
+interface Props extends Partial<React.ImgHTMLAttributes<HTMLImageElement>> {
   backgroundColor?: string
-  width?: number
-  height?: number
-  rounded?: boolean
+  width?: number | string
+  height?: number | string
+  rounded?: number
+  fit: 'contain' | 'cover'
+  'data-id'?: string | number
 }
 
 const defaultProps = {
   width: 100,
   height: 100,
+  fit: 'cover',
+  'data-id': null
 }
 
 const Image = ({
@@ -20,8 +24,10 @@ const Image = ({
   alt,
   rounded,
   backgroundColor,
+  fit,
+  'data-id': dataId,
   onClick,
-}: ImageProps & React.ImgHTMLAttributes<HTMLImageElement>) => {
+}: Props) => {
   return (
     <Img
       src={src}
@@ -29,8 +35,10 @@ const Image = ({
       width={width}
       rounded={rounded}
       height={height}
+      fit={fit}
       onClick={onClick}
       backgroundColor={backgroundColor}
+      data-id={dataId}
     />
   )
 }
@@ -38,19 +46,19 @@ const Image = ({
 const _PADDING_PERCENTAGE = 25
 const _DEFAULT_RADIUS = 7
 
-const Img = styled.img<ImageProps & React.ImgHTMLAttributes<HTMLImageElement>>`
-  ${({ backgroundColor, width, height, rounded, onClick }) => {
+const Img = styled.img<Props>`
+  ${({ backgroundColor, width, height, rounded, onClick, fit }) => {
     let styles: string = `
-      border-radius: ${rounded ? (width + height) / 2 : _DEFAULT_RADIUS}px;
-      width: ${width}px;
-      height: ${height}px;
-      object-fit: cover;
+      border-radius: ${rounded || _DEFAULT_RADIUS}px;
+      width: ${typeof width === 'string' ? width : `${width}px`};
+      height: ${typeof height === 'string' ? height : `${height}px`};
+      object-fit: ${fit};
     `
 
     if (backgroundColor?.length >= 1) {
       styles += `
         background-color: ${backgroundColor};
-        padding: ${(width * _PADDING_PERCENTAGE) / 100}px;
+        padding: ${((width as number) * _PADDING_PERCENTAGE) / 100}px;
       `
     }