+++ src/ui/layout/Section.tsx
@@ -1,12 +1,13 @@
 import React, { ReactNode, ReactElement } from 'react'
+import { css } from '@emotion/react'
 
 import Typography from '../atoms/Typography'
+import Button from '../atoms/Button'
 import Row from './Row'
-import { Button } from '..'
 
 interface Props {
   children: ReactNode | ReactElement | ReactElement[]
-  title: string
+  title: string | ReactElement
   description?: string
   rightChild?: string | ReactElement
 }
@@ -14,8 +15,10 @@ interface Props {
 function Section({ title, description, rightChild, children }: Props) {
   return (
     <section>
-      <Row align='space-between'>
-        <Typography type='title-3' size={16}>{title}</Typography>
+      <Row align='space-between' styles={sectionTitleStyles}>
+        {typeof title === 'string' ?
+          <Typography type='title-3' size={20}>{title}</Typography>
+          : title}
         <div>{rightChild}</div>
       </Row>
       <Typography type='description'>{description}</Typography>
@@ -24,6 +27,18 @@ function Section({ title, description, rightChild, children }: Props) {
     </section>
   )
 }
+
+const sectionTitleStyles = css`
+  @media screen and (min-width: 768px) {
+    h3 {
+      font-size: 30px;
+    }
+
+    button {
+      font-size: 20px;
+    }
+  }
+`
 
 interface RightChildAsButtonProps {
   onClick: () => void