+++ src/ui/ts/enums/InputSizesEnum.ts
@@ -1,5 +1,0 @@
-export enum InputSizesEnum {
-  small = 'small',
-  medium = 'medium',
-  large = 'large',
-}