+++ src/ui/ts/enums/UIElementStatusEnum.ts
@@ -1,0 +1,6 @@
+export enum UIElementStatusEnum {
+  Success = 'success',
+  Error = 'error',
+  Info = 'info',
+  Warning = 'warning',
+}